refactor: Added index-file to favorites and changed imports (#3375)
* refactor: Added index-file to favorites

* refactor: Changed imports to import from index-file
