+++ src/api/trips.ts
@@ -1,6 +1,6 @@
 import {TripPattern} from '@atb/sdk';
 import client from './client';
-import {Location} from '@atb/favorites/types';
+import {Location} from '@atb/favorites';
 import {AxiosRequestConfig} from 'axios';
 
 export default async function search(
