+++ src/nearby-stop-places/NearbyStopPlacesScreenComponent.tsx
@@ -4,8 +4,7 @@ import {SimpleDisappearingHeader} from '@atb/components/disappearing-header';
 import {ScreenReaderAnnouncement} from '@atb/components/screen-reader-announcement';
 import {LocationInputSectionItem, Section} from '@atb/components/sections';
 import {ThemeIcon} from '@atb/components/theme-icon';
-import {FavoriteChips} from '@atb/favorites';
-import {Location} from '@atb/favorites/types';
+import {FavoriteChips, Location} from '@atb/favorites';
 import {useGeolocationState} from '@atb/GeolocationContext';
 import {StopPlaces} from './components/StopPlaces';
 import {useNearestStopsData} from './use-nearest-stops-data';
