+++ src/stacks-hierarchy/Root_PurchaseTariffZonesSearchByTextScreen/VenueResults.tsx
@@ -1,7 +1,7 @@
 import {screenReaderPause} from '@atb/components/text';
 import {LocationIcon} from '@atb/components/location-icon';
 import {ThemeText} from '@atb/components/text';
-import {SearchLocation} from '@atb/favorites/types';
+import {SearchLocation} from '@atb/favorites';
 import {TariffZone} from '@atb/reference-data/types';
 import {getReferenceDataName} from '@atb/reference-data/utils';
 import {StyleSheet} from '@atb/theme';
