+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_DashboardStack/Dashboard_TripSearchScreen/use-trips-query.ts
@@ -1,4 +1,4 @@
-import {Location} from '@atb/favorites/types';
+import {Location} from '@atb/favorites';
 import {DateString, SearchTime} from '@atb/journey-date-picker';
 import {TripPattern} from '@atb/api/types/trips';
 import {
