+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_DashboardStack/types.ts
@@ -1,6 +1,6 @@
 import {TripPattern} from '@atb/api/types/trips';
 
-import {Location} from '@atb/favorites/types';
+import {Location} from '@atb/favorites';
 import {SearchTime} from '@atb/journey-date-picker';
 import {TransportModes} from '@atb/api/types/generated/journey_planner_v3_types';
 import {LanguageAndTextType} from '@atb/translations';
