+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_NearbyStack/Nearby_Root.tsx
@@ -12,7 +12,7 @@ import {
 } from '@atb/components/sections';
 import {ThemeIcon} from '@atb/components/theme-icon';
 import DeparturesList from '@atb/departure-list/DeparturesList';
-import {GeoLocation, Location} from '@atb/favorites/types';
+import {GeoLocation, Location} from '@atb/favorites';
 import {
   RequestPermissionFn,
   useGeolocationState,
