+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_NearbyStack/use-departure-data.ts
@@ -19,7 +19,7 @@ import {
 } from '@atb/api/departures/departure-group';
 import {ErrorType, getAxiosErrorType} from '@atb/api/utils';
 import {useFavorites} from '@atb/favorites';
-import {Location, UserFavoriteDepartures} from '@atb/favorites/types';
+import {Location, UserFavoriteDepartures} from '@atb/favorites';
 import {DeparturesRealtimeData} from '@atb/sdk';
 import {differenceInMinutes} from 'date-fns';
 import useInterval from '@atb/utils/use-interval';
