+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_ProfileStack/Profile_FavoriteListScreen.tsx
@@ -5,7 +5,7 @@ import {FullScreenHeader} from '@atb/components/screen-header';
 import * as Sections from '@atb/components/sections';
 import {ThemeIcon} from '@atb/components/theme-icon';
 import {useFavorites} from '@atb/favorites';
-import {StoredLocationFavorite} from '@atb/favorites/types';
+import {StoredLocationFavorite} from '@atb/favorites';
 import {StyleSheet, Theme} from '@atb/theme';
 import {FavoriteListTexts, useTranslation} from '@atb/translations';
 import React from 'react';
